requirements.txt
numpy
pandas
scikit-image
torch
torchvision
pyyaml
tqdm
matplotlib